+++ CRM_Frontend/src/app/view/accountDetail/errorPage/errorPage.html
@@ -1,1 +1,1 @@
-<h1 class="text-danger text_font" id = "loadDateError">You don't have authentication to access</h1>
+<h1 class="text-danger text_font" id = "loadDateError">{{erroeMessage}}</h1>
