+++ CRM_Frontend/src/app/view/accountDetail/errorPage/errorPage.ts
@@ -1,10 +1,11 @@
-import { Component } from '@angular/core';
+import {Component, Input} from '@angular/core';
 
 @Component({
   selector: 'app-error-page',
   templateUrl: './errorPage.html',
   styleUrls: ['./errorPage.css']
 })
-export class ErrorPage {
-
+export class ErrorPage
+{
+  @Input() erroeMessage:any;
 }
