+++ CRM_Frontend/src/app/view/accountDetail/password/password.component.html
@@ -1,0 +1,31 @@
+<h3 class="card-title text-center border-bottom pb-4 text-white text_font">Reset Password</h3>
+<div style="min-height: 260px">
+    <div class="row m-3 mt-5" *ngIf="this.isMyAccount == true">
+        <div class="col-3 text-end">
+            <label class="text_font text-white">Original</label>
+        </div>
+        <div class="col-7 text-end">
+            <input class="form-control text_font" type="password" value="" id="originalPassword">
+        </div>
+    </div>
+    <div class="row m-3 mt-5">
+        <div class="col-3 text-end">
+            <label class="text_font text-white">New</label>
+        </div>
+        <div class="col-7 text-end">
+            <input class="form-control text_font" type="password" value="" id="newPassword">
+        </div>
+    </div>
+    <div class="row m-3">
+        <div class="col-3 text-end">
+            <label class="text_font text-white">Confirm</label>
+        </div>
+        <div class="col-7 text-end">
+            <input class="form-control text_font" type="password" value="" id="confirmPassword">
+        </div>
+    </div>
+    <div id="errorMessage"></div>
+</div>
+<div class="col-sm-12 text-end mt-4 pt-2">
+    <button class="btn btn-primary mb-3 btn-lg py-1 px-5 ms-4" (click)="resetPassword()">Update</button>
+</div>
