+++ CRM_Frontend/src/app/view/nav/footer/footer.component.html
@@ -1,11 +1,8 @@
 <!-- Footer -->
 <footer class="text-center text-lg-start text-white cardBackGround border-success border-top border-bottom">
-  <!-- Copyright -->
   <div class="text-center" >
     © 2023 Copyright:
     <a class="text-reset fw-bold" href="https://mdbootstrap.com/">MDBootstrap.com</a>
   </div>
-  <!-- Copyright -->
 </footer>
 <!-- Footer -->
-
