+++ CRM_Frontend/src/styles.css
@@ -14,6 +14,7 @@ footer{
   min-height: 90px;
   float: left;
   margin-top:0 !important;
+  z-index:99999;
 }
 
 .cardBackGround
